Bucket/*: more consistent error handling
